sprint 3

+++ app/src/main/java/com/devfiveurjc/agendaly/fragments/TaskInfoFragment.java
@@ -1,11 +1,13 @@
 package com.devfiveurjc.agendaly.fragments;
 
+import android.app.AlertDialog;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.fragment.NavHostFragment;
 
+import android.view.ContextThemeWrapper;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -40,7 +42,7 @@ public class TaskInfoFragment extends Fragment {
         title.setText(task.getTitle());
         description.setText(task.getDescription());
         date.setText(task.getDate().toString());
-        status.setText((task.isCheck()) ? "Completed" : "Uncompleted");
+        status.setText((task.isCheck()) ? R.string.completed : R.string.uncompleted);
         return view;
     }
 
@@ -48,17 +50,29 @@ public class TaskInfoFragment extends Fragment {
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         this.binding.taskInfoDeleteButton.setOnClickListener(v -> {
-            // TODO: confirmation popup
-            CRUDTask.deleteTask(this.taskId);
-            NavHostFragment.findNavController(this)
-                    .navigate(R.id.action_TaskInfoFragment_to_TaskListFragment);
+            this.deleteCautionMessage();
         });
         this.binding.taskInfoEditButton.setOnClickListener(v -> {
             Bundle bundle = new Bundle();
-            bundle.putInt("taskId", taskId);
+            bundle.putInt("taskId", this.taskId);
             NavHostFragment.findNavController(this)
-                    .navigate(R.id.action_TaskInfoFragment_to_TaskEditFragment,bundle);
+                    .navigate(R.id.action_TaskInfoFragment_to_TaskEditFragment, bundle);
         });
+    }
+
+    private void deleteCautionMessage() {
+        ContextThemeWrapper newContext = new ContextThemeWrapper(this.getContext(), R.style.Theme_Agendaly_AlertDialog);
+        AlertDialog.Builder alert = new AlertDialog.Builder(newContext);
+        alert.setMessage(getString(R.string.caution_text_delete) + getString(R.string.caution_advise))
+            .setCancelable(false)
+            .setTitle(R.string.caution)
+            .setPositiveButton(R.string.confirm, (dialogInterface, i) -> {
+                CRUDTask.deleteTask(this.taskId);
+                NavHostFragment.findNavController(this)
+                    .navigate(R.id.action_TaskInfoFragment_to_TaskListFragment);
+            })
+            .setNegativeButton(R.string.goback, (dialogInterface, i) -> dialogInterface.cancel());
+        alert.show();
     }
 
     @Override
